requirements.txt
torch
torchvision
numpy
matplotlib
pillow
open3d
